myenv/requirements.txt
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
flask==2.3.3
python-dotenv==1.0.0
numpy==1.24.3 